add_slide {officer}	R Documentation

Add a slide

Description

Add a slide into a pptx presentation.

Usage

add_slide(x, layout = NULL, master = NULL, ..., .dots = NULL)

Arguments

x
an rpptx object.

layout
slide layout name to use. Can be ommited of a default layout is set via layout_default().

master
master layout name where layout is located. Only required in case of several masters if layout is not unique.

...
Key-value pairs of the form "short form location" = object passed to phs_with. See section "Short forms" in phs_with for details, available short forms and examples.

.dots
List of key-value pairs of the form list("short form location" = object). Alternative to .... See phs_with for details.

See Also

print.rpptx(), read_pptx(), layout_summary(), plot_layout_properties(), ph_with(), phs_with(), layout_default()

Other functions to manipulate slides: move_slide(), on_slide(), remove_slide(), set_notes()

Examples

x <- read_pptx()

layout_summary(x) # available layouts

x <- add_slide(x, layout = "Two Content")

x <- layout_default(x, "Title Slide") # set default layout for `add_slide()`
x <- add_slide(x) # uses default layout

# use `...` to fill placeholders when adding slide
x <- add_slide(
  x,
  layout = "Two Content",
  `Title 1` = "A title",
  dt = "Jan. 26, 2025",
  `body[2]` = "Body 2",
  left = "Left side",
  `6` = "Footer"
)

[Package officer version 0.7.3 Index]