as.matrix.rpptx {officer}	R Documentation

PowerPoint table to matrix

Description

Convert the data in an a 'PowerPoint' table to a matrix or all data to a list of matrices.

Usage

## S3 method for class 'rpptx'
as.matrix(
  x,
  ...,
  slide_id = NA_integer_,
  id = NA_character_,
  span = c(NA_character_, "fill")
)

Arguments

x
The rpptx object to convert (as created by read_pptx())

...
Ignored

slide_id
The slide number to load from (NA indicates first slide with a table, NULL indicates all slides and all tables)

id
The table ID to load from (ignored it is.null(slide_id), NA indicates to load the first table from the slide_id)

span
How should col_span/row_span values be handled? NA means to leave the value as NA, and "fill" means to fill matrix cells with the value.

Value

A matrix with the data, or if slide_id=NULL, a list of matrices

Examples

library(officer)
pptx_file <- system.file(package="officer", "doc_examples", "example.pptx")
z <- read_pptx(pptx_file)
as.matrix(z, slide_id = NULL)

[Package officer version 0.7.3 Index]