caption block

Create a representation of a caption that can be used for cross reference. The caption can also be an auto numbered paragraph.

block_caption(label, style, id, autonum = NULL)

Arguments

label
a scalar character representing label to display

style
paragraph style name

id
cross reference identifier

autonum
an object generated with function run_autonum

See also

Examples

block_caption("a caption", style = "Normal", id = "caption_id")

#> caption [anto-num on]: a caption

block_caption("a caption",
  style = "Normal", id = "caption_id",
  autonum = run_autonum()
)

#> caption [antonum off]: a caption