block_pour_docx {officer}	R Documentation

Pour external Word document in the current document

Description

Pour the content of a docx file in the resulting docx generated by the main R Markdown document.

Usage

block_pour_docx(file)

Arguments

file
external docx file path

See Also

Other block functions for reporting: block_caption(), block_list(), block_section(), block_table(), block_toc(), fpar(), plot_instr(), unordered_list()

Examples

library(officer)
docx <- tempfile(fileext = ".docx")
doc <- read_docx()
doc <- body_add(doc, iris[1:20,], style = "table_template")
print(doc, target = docx)

target <- tempfile(fileext = ".docx")
doc_1 <- read_docx()
doc_1 <- body_add(doc_1, block_pour_docx(docx))
print(doc_1, target = target)

[Package officer version 0.4.1 Index]