block_toc {officer}	R Documentation

Table of content

Description

Create a representation of a table of content.

Usage

block_toc(level = 3, style = NULL, seq_id = NULL, separator = ";")

Arguments

level
max title level of the table

style
optional. If not NULL, its value is used as style in the document that will be used to build entries of the TOC.

seq_id
optional. If not NULL, its value is used as sequence identifier in the document that will be used to build entries of the TOC. See also run_autonum() to specify a sequence identifier.

separator
optional. Some configurations need "," (i.e. from Canada) separator instead of ";"

See Also

Other block functions for reporting: block_caption(), block_list(), block_pour_docx(), block_section(), block_table(), fpar(), plot_instr(), unordered_list()

Examples

block_toc(level = 2)
block_toc(style = "Table Caption")

[Package officer version 0.4.1 Index]