body_add_table {officer}	R Documentation

add table

Description

add a table into an rdocx object

Usage

body_add_table(
  x,
  value,
  style = NULL,
  pos = "after",
  header = TRUE,
  alignment = NULL,
  align_table = "center",
  stylenames = table_stylenames(),
  first_row = TRUE,
  first_column = FALSE,
  last_row = FALSE,
  last_column = FALSE,
  no_hband = FALSE,
  no_vband = TRUE
)

Arguments

x
a docx device

value
a data.frame to add as a table

style
table style

pos
where to add the new element relative to the cursor, one of after", "before", "on".

header
display header if TRUE

alignment
columns alignement, argument length must match with columns length, values must be "l" (left), "r" (right) or "c" (center).

align_table
table alignment within document, value must be "left", "center" or "right"

stylenames
columns styles defined by table_stylenames()

first_row
Specifies that the first column conditional formatting should be applied. Details for this and other conditional formatting options can be found at http://officeopenxml.com/WPtblLook.php.

first_column
Specifies that the first column conditional formatting should be applied.

last_row
Specifies that the first column conditional formatting should be applied.

last_column
Specifies that the first column conditional formatting should be applied.

no_hband
Specifies that the first column conditional formatting should be applied.

no_vband
Specifies that the first column conditional formatting should be applied.

See Also

Other functions for adding content: body_add_blocks(), body_add_break(), body_add_caption(), body_add_docx(), body_add_fpar(), body_add_gg(), body_add_img(), body_add_par(), body_add_plot(), body_add_toc()

Examples

doc <- read_docx()
doc <- body_add_table(doc, iris, style = "table_template")

print(doc, target = tempfile(fileext = ".docx") )

[Package officer version 0.4.1 Index]