body_add_toc {officer}	R Documentation

add table of content

Description

add a table of content into an rdocx object. The TOC will be generated by Word, if the document is not edited with Word (i.e. Libre Office) the TOC will not be generated.

Usage

body_add_toc(x, level = 3, pos = "after", style = NULL, separator = ";")

Arguments

x
an rdocx object

level
max title level of the table

pos
where to add the new element relative to the cursor, one of "after", "before", "on".

style
optional. style in the document that will be used to build entries of the TOC.

separator
optional. Some configurations need "," (i.e. from Canada) separator instead of ";"

See Also

Other functions for adding content: body_add_blocks(), body_add_break(), body_add_caption(), body_add_docx(), body_add_fpar(), body_add_gg(), body_add_img(), body_add_par(), body_add_plot(), body_add_table()

Examples

doc <- read_docx()
doc <- body_add_toc(doc)

print(doc, target = tempfile(fileext = ".docx") )

[Package officer version 0.4.1 Index]