add images into an rdocx object

reference images into a Word document. This function is to be used with wml_link_images.

Images need to be referenced into the Word document, this will generate unique identifiers that need to be known to link these images with their corresponding xml code (wml).

docx_reference_img(x, src)

Arguments

x
an rdocx object

src
a vector of character containing image filenames.

See also