fortify_location {officer}	R Documentation

eval a location on the current slide

Description

Eval a shape location against the current slide. This function is to be used to add custom openxml code. A list is returned, it contains informations width, height, left and top positions and other informations necessary to add a content on a slide.

Usage

fortify_location(x, doc, ...)

Arguments

x
a location for a placeholder.

doc
an rpptx object

...
unused arguments

See Also

ph_location, ph_with

Other functions for officer extensions: docx_reference_img(), get_reference_value(), opts_current_table(), to_html(), to_pml(), to_wml(), wml_link_images()

Examples

doc <- read_pptx()
doc <- add_slide(doc, layout = "Title and Content",
  master = "Office Theme")
fortify_location(ph_location_fullsize(), doc)

[Package officer version 0.4.1 Index]