fp_tab {officer}	R Documentation

Tabulation mark properties object

Description

create a tabulation mark properties setting object for Word or RTF. Results can be used as arguments of fp_tabs().

Once tabulation marks settings are defined, tabulation marks can be added with run_tab() inside a call to fpar() or with \t within 'flextable' content.

Usage

fp_tab(pos, style = "decimal")

Arguments

pos
Specifies the position of the tab stop (in inches).

style
style of the tab. Possible values are: "decimal", "left", "right" or "center".

See Also

Other functions for defining formatting properties: fp_border(), fp_cell(), fp_par(), fp_tabs(), fp_text()

Examples

fp_tab(pos = 0.4, style = "decimal")
fp_tab(pos = 1, style = "right")

[Package officer version 0.7.3 Index]