fp_tabs {officer}	R Documentation

Tabs properties object

Description

create a set of tabulation mark properties object for Word or RTF. Results can be used as arguments tabs of fp_par() and will only have effects in Word or RTF outputs.

Once a set of tabulation marks settings is defined, tabulation marks can be added with run_tab() inside a call to fpar() or with \t within 'flextable' content.

Usage

fp_tabs(...)

Arguments

...
fp_tab objects

See Also

Other functions for defining formatting properties: fp_border(), fp_cell(), fp_par(), fp_tab(), fp_text()

Examples

z <- fp_tabs(
  fp_tab(pos = 0.4, style = "decimal"),
  fp_tab(pos = 1, style = "decimal")
)
fpar(
  run_tab(), ftext("88."),
  run_tab(), ftext("987.45"),
  fp_p = fp_par(
    tabs = z
  )
)

[Package officer version 0.7.3 Index]