Text formatting properties

Create a fp_text object that describes text formatting properties.

fp_text(color = "black", font.size = 10, bold = FALSE, italic = FALSE,
  underlined = FALSE, font.family = "Arial", vertical.align = "baseline",
  shading.color = "transparent")

# S3 method for fp_text
format(x, type = "wml", ...)

# S3 method for fp_text
print(x, ...)

# S3 method for fp_text
as.data.frame(x, ...)

# S3 method for fp_text
update(object, color, font.size, bold = FALSE,
  italic = FALSE, underlined = FALSE, font.family, vertical.align,
  shading.color, ...)

Arguments

color

font color - a single character value specifying a valid color (e.g. "#000000" or "black").

font.size

font size (in point) - 0 or positive integer value.

bold

is bold

italic

is italic

underlined

is underlined

font.family

single character value specifying font name.

vertical.align

single character value specifying font vertical alignments. Expected value is one of the following : default 'baseline' or 'subscript' or 'superscript'

shading.color

shading color - a single character value specifying a valid color (e.g. "#000000" or "black").

x

fp_text object

type

output type - one of 'wml', 'pml', 'html'.

...

further arguments - not used

object

fp_text object to modify

format

format type, wml for MS word, pml for MS PowerPoint and html.

Value

a fp_text object

Examples

print( fp_text (color="red", font.size = 12) )

#>                     values
#> size                    12
#> italic               FALSE
#> bold                 FALSE
#> underlined           FALSE
#> color                  red
#> shading        transparent
#> fontname             Arial
#> vertical_align    baseline