get_reference_value {officer}	R Documentation

Get the document being used as a template

Description

Get filename of the document being used as a template in an R Markdown document rendered as HTML, PowerPoint presentation or Word document. It requires packages rmarkdown >= 1.10.14 and knitr.

Usage

get_reference_value(format = NULL)

Arguments

format
document format, one of 'pptx', 'docx' or 'html'

Value

a name file

See Also

Other functions for officer extensions: docx_reference_img(), fortify_location(), opts_current_table(), to_html(), to_pml(), to_wml(), wml_link_images()

[Package officer version 0.4.1 Index]