layout_properties {officer}	R Documentation

Slide layout properties

Description

Detailed information about the placeholders on the slide layouts (label, position, etc.). See Value section below for more info.

Usage

layout_properties(x, layout = NULL, master = NULL)

Arguments

x
an rpptx object

layout
slide layout name. If NULL, returns all layouts.

master
master layout name where layout is located. If NULL, returns all masters.

Value

Returns a data frame with one row per placeholder and the following columns:

master_name: Name of master (a .pptx file may have more than one)

name: Name of layout

type: Placeholder type

type_idx: Running index for phs of the same type. Ordering by ph position (top -> bottom, left -> right)

id: A unique placeholder id (assigned by PowerPoint automatically, starts at 2, potentially non-consecutive)

ph_label: Placeholder label (can be set by the user in PowerPoint)

ph: Placholder XML fragment (usually not needed)

offx,offy: placeholder's distance from left and top edge (in inch)

cx,cy: width and height of placeholder (in inch)

rotation: rotation in degrees

fld_id is generally stored as a hexadecimal or GUID value

fld_type: a unique identifier for a particular field

See Also

Other functions for reading presentation information: annotate_base(), color_scheme(), doc_properties(), layout_summary(), length.rpptx(), plot_layout_properties(), slide_size(), slide_summary()

Examples

library(officer)

x <- read_pptx()
layout_properties(x = x, layout = "Title Slide", master = "Office Theme")
layout_properties(x = x, master = "Office Theme")
layout_properties(x = x, layout = "Two Content")
layout_properties(x = x)

[Package officer version 0.7.3 Index]