length.rdocx {officer}	R Documentation

number of blocks inside an rdocx object

Description

return the number of blocks inside an rdocx object. This number also include the default section definition of a Word document - default Word section is an uninvisible element.

Usage

## S3 method for class 'rdocx'
length(x)

Arguments

x
an rdocx object

See Also

Other functions for Word document informations: doc_properties(), docx_bookmarks(), docx_dim(), set_doc_properties(), styles_info()

Examples

# how many elements are there in an new document produced
# with the default template.
length( read_docx() )

[Package officer version 0.4.1 Index]