Location of a named placeholder for notes

The function will use the label of a placeholder to find the corresponding location in the slide notes.

Usage

notes_location_label(ph_label, ...)

Arguments

ph_label

placeholder label of the used notes master

...

unused arguments