ph_add_par {officer}	R Documentation

append paragraph

Description

append a new empty paragraph in a placeholder. The function let you add a new empty paragraph to an existing content in an exiisting shape, existing paragraphs will be preserved.

Usage

ph_add_par(x, type = "body", id = 1, id_chr = NULL, level = 1, ph_label = NULL)

Arguments

x
an rpptx object

type
placeholder type

id
placeholder index (integer) for a duplicated type. This is to be used when a placeholder type is not unique in the layout of the current slide, e.g. two placeholders with type 'body'. To add onto the first, use id = 1 and id = 2 for the second one. Values can be read from slide_summary.

id_chr
deprecated.

level
paragraph level

ph_label
label associated to the placeholder. Use column ph_label of result returned by slide_summary.

Usage

If your goal is to add formatted text in a new shape, use ph_with with a block_list instead of this function.

Note

This function will be deprecated in a next release because it is not efficient and make users write complex code. Use instead fpar() to build formatted paragraphs.

[Package officer version 0.4.1 Index]