Hyperlink a placeholder

Add hyperlink to a placeholder in the current slide.

Usage

ph_hyperlink(x, type = "body", id = 1, id_chr = NULL, ph_label = NULL, href)

Arguments

x

an rpptx object

type

placeholder type

id

placeholder index (integer) for a duplicated type. This is to be used when a placeholder type is not unique in the layout of the current slide, e.g. two placeholders with type 'body'. To add onto the first, use id = 1 and id = 2 for the second one. Values can be read from slide_summary.

id_chr

deprecated.

ph_label

label associated to the placeholder. Use column ph_label of result returned by slide_summary. If used, type and id are ignored.

href

hyperlink (do not forget http or https prefix)

See also

Examples

fileout <- tempfile(fileext = ".pptx")
loc_manual <- ph_location(bg = "red", newlabel = "mytitle")
doc <- read_pptx()
doc <- add_slide(doc)
doc <- ph_with(x = doc, "Un titre 1", location = loc_manual)
slide_summary(doc) # read column ph_label here
#>   type                                   id ph_label offx offy cx cy rotation
#> 1 body 94ce5686-20f1-45b8-85db-7f9a1a020a13  mytitle    1    1  4  3       NA
#>   fld_id fld_type       text
#> 1   <NA>     <NA> Un titre 1
doc <- ph_hyperlink(
  x = doc, ph_label = "mytitle",
  href = "https://cran.r-project.org"
)

print(doc, target = fileout)