create a location for a placeholder

The function will return a list that complies with expected format for argument location of functions ph_with_* and ph_with methods.

ph_location(left = 1, top = 1, width = 4, height = 3, label = "",
  ph = "<p:ph/>", bg = NULL, rotation = NULL, ...)

Arguments

left, top, width, height
place holder coordinates in inches.

label
a label for the placeholder. See section details.

ph
openxml string value for tag. This is not meant to be used by end users (unless they understand openxml placeholder and p:ph tag).

bg
background color

rotation
rotation angle

...
unused arguments

Details

The location of the bounding box associated to a placeholder within a presentation slide is specified with the left top coordinate, the width and the height. These are defined in inches:

left

left coordinate of the bounding box

top

top coordinate of the bounding box

width

width of the bounding box

height

height of the bounding box

In addition to these attributes, there is attribute ph_label associated with the shape (shapes, text boxes, images and other objects will be identified with that label in the Selection Pane of PowerPoint). This label can then be reused by other functions such as ph_add_fpar, or ph_add_text.

See also

Examples

library(magrittr)
read_pptx() %>%
  add_slide(layout = "Title and Content", master = "Office Theme") %>%
  ph_with("Hello world",
    location = ph_location(width = 4, height = 3, label = "hello") ) %>%
  print(target = tempfile(fileext = ".pptx") )

#> [1] "/private/var/folders/08/2qdvv0q95wn52xy6mxgj340r0000gn/T/RtmpXIYvn5/filef90c40efa8e8.pptx"