add object into a new shape

add object into a new shape in a slide.

ph_with(x, value, ...)

# S3 method for character
ph_with(x, value, location, ...)

# S3 method for block_list
ph_with(x, value, location, ...)

# S3 method for data.frame
ph_with(x, value, location, header = TRUE,
  first_row = TRUE, first_column = FALSE, last_row = FALSE,
  last_column = FALSE, ...)

# S3 method for gg
ph_with(x, value, location, ...)

Arguments

x
a pptx device

value
object to add as a new shape.

...
Arguments to be passed to methods

location
a placeholder location object. This is a convenient argument that can replace usage of arguments type and index. See section see also.

header
display header if TRUE

first_row, last_row, first_column, last_column
logical for PowerPoint table options

with character

When value is a character vector, each value will be added as a paragraph.

with block_list

When value is a block_list object, each value will be added as a paragraph.

with data.frame

When value is a data.frame, a simple table is added, use package flextable instead for more advanced formattings.

with a ggplot object

When value is a ggplot object, a raster plot is produced and added, use package rvg instead for more advanced graphical features.

See also

Examples

library(magrittr)

fileout <- tempfile(fileext = ".pptx")
doc <- read_pptx()
doc <- add_slide(doc, layout = "Title and Content",
  master = "Office Theme")

if( require("ggplot2") ){
  gg_plot <- ggplot(data = iris ) +
    geom_point(mapping = aes(Sepal.Length, Petal.Length),
      size = 3) +
    theme_minimal()
  doc <- ph_with(x = doc, value = gg_plot,
                 location = ph_location_fullsize() )
}

doc <- ph_with(x = doc, value = c("Un titre", "Deux titre"),
               location = ph_location_left() )
doc <- ph_with(x = doc, value = iris[1:4, 3:5],
               location = ph_location_right() )

print(doc, target = fileout )

#> [1] "/private/var/folders/08/2qdvv0q95wn52xy6mxgj340r0000gn/T/RtmpCZRqgK/file253b5bb95dac.pptx"