section properties

A section is a grouping of blocks (ie. paragraphs and tables) that have a set of properties that define pages on which the text will appear.

A Section properties object stores information about page composition, such as page size, page orientation, borders and margins.

prop_section(
  page_size = NULL,
  page_margins = NULL,
  type = NULL,
  section_columns = NULL
)

Arguments

page_size
page dimensions, an object generated with function page_size.

page_margins
page margins, an object generated with function page_mar.

type
Section type. It defines how the contents of the section will be placed relative to the previous section. Available types are "continuous" (begins the section on the next paragraph), "evenPage" (begins on the next even-numbered page), "nextColumn" (begins on the next column on the page), "nextPage" (begins on the following page), "oddPage" (begins on the next odd-numbered page).

section_columns
section columns, an object generated with function section_columns.

See also

Examples

prop_section(
  page_size = page_size(orient = "landscape"),
  page_margins = page_mar(top = 2),
  type = "continuous")

#> $type
#> [1] "continuous"
#> 
#> $page_size
#> $width
#> [1] 11.69291
#> 
#> $height
#> [1] 8.267717
#> 
#> $orient
#> [1] "landscape"
#> 
#> attr(,"class")
#> [1] "page_size"
#> 
#> $page_margins
#> $header
#> [1] 0.5
#> 
#> $bottom
#> [1] 1
#> 
#> $top
#> [1] 2
#> 
#> $right
#> [1] 1
#> 
#> $left
#> [1] 1
#> 
#> $footer
#> [1] 0.5
#> 
#> $gutter
#> [1] 0.5
#> 
#> attr(,"class")
#> [1] "page_mar"
#> 
#> attr(,"class")
#> [1] "prop_section" "prop"