open a connexion to a 'Word' file

read and import a docx file as an R object representing the document.

read_docx(path = NULL)

# S3 method for rdocx
print(x, target = NULL, ...)

# S3 method for rdocx
length(x)

Arguments

path

path to the docx file to use a base document.

x

a rdocx object

target

path to the docx file to write

...

unused

Examples

# create a rdocx object with default template ---
read_docx()

#> rdocx document with 2 element(s)
#> Available styles are:
#>    style_type             style_name
#> 1   paragraph                 Normal
#> 2   paragraph              heading 1
#> 3   paragraph              heading 2
#> 4   paragraph              heading 3
#> 5   character Default Paragraph Font
#> 6       table           Normal Table
#> 7   numbering                No List
#> 8   character                 strong
#> 9   paragraph               centered
#> 10      table         table_template
#> 11      table    Light List Accent 2
#> 12  character            Titre 1 Car
#> 13  character            Titre 2 Car
#> 14  character            Titre 3 Car
#> 15  paragraph          graphic title
#> 16  paragraph            table title
#> 17      table     Table Professional
#> 18  paragraph                  toc 1
#> 19  paragraph                  toc 2
#> 20  paragraph           Balloon Text
#> 21  character    Texte de bulles Car

# write a rdocx object in a docx file ----
if( require(magrittr) && has_zip() ){
  read_docx() %>% print(target = "out.docx")
  # full path of produced file is returned
  print(.Last.value)
}

#> 
#> ── officer 0.1.0.9000: NOTE
#> 
#>   Build ID:   officer_0.1.0.9000.tar.gz-ce1d6832db81441f9519a65071a413ed
#>   Platform:   Windows Server 2008 R2 SP1, R-devel, 32/64 bit
#>   Submitted:  5.9s ago
#>   Build time: 9m 41.5s
#> 
#> 
#> ── R CMD check results ───────────────────────────────── officer 0.1.0.9000 ────
#> 
#> ❯ checking CRAN incoming feasibility ... NOTE
#>   Maintainer: 'David Gohel <david.gohel@ardata.fr>'
#>   
#>   Version contains large components (0.1.0.9000)
#> 
#> ── 0 errors ✔ | 0 warnings ✔ | 1 note ✖
#> 
#> ── officer 0.1.0.9000: WARNING
#> 
#>   Build ID:   officer_0.1.0.9000.tar.gz-65e51a1aecbd460cbb6919806c1401fc
#>   Platform:   Ubuntu Linux 16.04 LTS, R-release, GCC
#>   Submitted:  5.9s ago
#>   Build time: 6m 14s
#> 
#> 
#> ── R CMD check results ───────────────────────────────── officer 0.1.0.9000 ────
#> 
#> ❯ checking re-building of vignette outputs ... WARNING
#>   Error in re-building vignettes:
#>     ...
#>   sh: 1: /usr/bin/zip: not found
#>   Quitting from lines 112-114 (powerpoint.Rmd) 
#>   Error: processing vignette 'powerpoint.Rmd' failed with diagnostics:
#>   could not zip package '/tmp/RtmpkJP1iV/file64a2da7a895' in file '/home/docker/officer.Rcheck/vign_test/officer/vignettes/assets/pptx/first_example.pptx'
#>   Execution halted
#> 
#> ❯ checking CRAN incoming feasibility ... NOTE
#>   Maintainer: ‘David Gohel <david.gohel@ardata.fr>’
#>   
#>   Version contains large components (0.1.0.9000)
#> 
#> ── 0 errors ✔ | 1 warning ✖ | 1 note ✖
#> 
#> ── officer 0.1.0.9000: ERROR
#> 
#>   Build ID:   officer_0.1.0.9000.tar.gz-8ac39cf7068249b59cfbfc76fc38be8d
#>   Platform:   Fedora Linux, R-devel, clang, gfortran
#>   Submitted:  5.9s ago
#>   Build time: 9m 40.9s
#> 
#> 
#> ── R CMD check results ───────────────────────────────── officer 0.1.0.9000 ────
#> 
#> ❯ checking examples ... ERROR
#>   Running examples in ‘officer-Ex.R’ failed
#>   The error most likely occurred in:
#>   
#>   > base::assign(".ptime", proc.time(), pos = "CheckExEnv")
#>   > ### Name: body_add_gg
#>   > ### Title: add ggplot
#>   > ### Aliases: body_add_gg
#>   > 
#>   > ### ** Examples
#>   > 
#>   > library(ggplot2)
#>   > 
#>   > doc <- read_docx()
#>   > 
#>   > gg_plot <- ggplot(data = iris ) +
#>   +   geom_point(mapping = aes(Sepal.Length, Petal.Length))
#>   > 
#>   > doc <- body_add_gg(doc, value = gg_plot, style = "centered" )
#>   Error in grid.Call.graphics(C_text, as.graphicsAnnot(x$label), x$x, x$y,  : 
#>     X11 font -adobe-helvetica-%s-%s-*-*-%d-*-*-*-*-*-*-*, face 1 at size 12 could not be loaded
#>   Error in grid.Call(C_stringMetric, as.graphicsAnnot(x$label)) : 
#>     X11 font -adobe-helvetica-%s-%s-*-*-%d-*-*-*-*-*-*-*, face 1 at size 12 could not be loaded
#>   Calls: body_add_gg ... titleGrob -> descentDetails -> descentDetails.text -> grid.Call
#>   Execution halted
#> 
#> ❯ checking tests ...
#>     Running ‘testthat.R’ [15s/15s]
#>    ERROR
#>   Running the tests in ‘tests/testthat.R’ failed.
#>   Last 13 lines of output:
#>     27: grid.Call(C_stringMetric, as.graphicsAnnot(x$label))
#>     
#>     sh: : command not found
#>     sh: : command not found
#>     sh: : command not found
#>     sh: : command not found
#>     sh: : command not found
#>     sh: : command not found
#>     sh: : command not found
#>     testthat results ================================================================
#>     OK: 307 SKIPPED: 7 FAILED: 1
#>     1. Error: ggplot add (@test-docx-add.R#79) 
#>     
#>     Error: testthat unit tests failed
#>     Execution halted
#> 
#> ❯ checking re-building of vignette outputs ... WARNING
#>   Error in re-building vignettes:
#>     ...
#>   sh: : command not found
#>   Quitting from lines 112-114 (powerpoint.Rmd) 
#>   Error: processing vignette 'powerpoint.Rmd' failed with diagnostics:
#>   A zip application must be available to R. You can set its path with Sys.setenv("R_ZIPCMD" = "path/to/zip")
#>   Execution halted
#> 
#> ❯ checking CRAN incoming feasibility ... NOTE
#>   Maintainer: ‘David Gohel <david.gohel@ardata.fr>’
#>   
#>   Version contains large components (0.1.0.9000)
#> 
#> ── 2 errors ✖ | 1 warning ✖ | 1 note ✖
#> 
#> ── officer 0.1.0.9000: IN-PROGRESS
#> 
#>   Build ID:   officer_0.1.0.9000.tar.gz-d21ddfbef31f47bd826e34d2ffc92410
#>   Platform:   Debian Linux, R-devel, GCC ASAN/UBSAN
#>   Submitted:  5.9s ago
#> 

# how many element are there in the document ----
length( read_docx() )

#> [1] 2