read_xlsx {officer}	R Documentation

open a connexion to an 'Excel' file

Description

read and import an xlsx file as an R object representing the document. This function is experimental.

Usage

read_xlsx(path = NULL)

## S3 method for class 'rxlsx'
length(x)

## S3 method for class 'rxlsx'
print(x, target = NULL, ...)

Arguments

path
path to the xlsx file to use as base document.

x
an rxlsx object

target
path to the xlsx file to write

...
unused

Examples

read_xlsx()
x <- read_xlsx()
print(x, target = tempfile(fileext = ".xlsx"))

[Package officer version 0.4.1 Index]