run_bookmark {officer}	R Documentation

bookmark for Word

Description

Add a bookmark on a run object.

Usage

run_bookmark(bkm, run)

Arguments

bkm
bookmark id to associate with run. Value can only be made of alpha numeric characters, '-' and '_'.

run
a run object, made with a call to one of the "run functions for reporting".

usage

You can use this function in conjunction with fpar to create paragraphs consisting of differently formatted text parts. You can also use this function as an r chunk in an R Markdown document made with package officedown.

See Also

Other run functions for reporting: external_img(), ftext(), hyperlink_ftext(), run_autonum(), run_columnbreak(), run_footnoteref(), run_footnote(), run_linebreak(), run_pagebreak(), run_reference(), run_word_field()

Examples

ft <- fp_text(font.size = 12, bold = TRUE)
run_bookmark("par1", ftext("some text", ft))

[Package officer version 0.4.1 Index]