run_footnoteref {officer}	R Documentation

Word footnote reference

Description

Wraps a footnote reference in an object that can then be inserted as a run/chunk with fpar() or within an R Markdown document.

Usage

run_footnoteref(prop = NULL)

Arguments

prop
formatting text properties returned by fp_text_lite() or fp_text(). It also can be NULL in which case, no formatting is defined (the default is applied).

See Also

Other run functions for reporting: external_img(), ftext(), hyperlink_ftext(), run_autonum(), run_bookmark(), run_columnbreak(), run_footnote(), run_linebreak(), run_pagebreak(), run_reference(), run_word_field()

Examples

run_footnoteref()
to_wml(run_footnoteref())

[Package officer version 0.4.1 Index]