run_wordtext {officer}	R Documentation

Word chunk of text with a style

Description

Format a chunk of text associated with a 'Word' character style. The style is defined with its unique identifer.

Usage

run_wordtext(text, style_id = NULL)

Arguments

text
text value, a single character value

style_id
'Word' unique style identifier associated with the style to use.

See Also

ftext()

Other run functions for reporting: external_img(), floating_external_img(), ftext(), hyperlink_ftext(), run_autonum(), run_bookmark(), run_columnbreak(), run_comment(), run_footnote(), run_footnoteref(), run_linebreak(), run_pagebreak(), run_reference(), run_tab(), run_word_field()

Examples

run1 <- run_wordtext("hello", "DefaultParagraphFont")
paragraph <- fpar(run1)

x <- read_docx()
x <- body_add_fpar(x, paragraph)
print(x, target = tempfile(fileext = ".docx"))

[Package officer version 0.7.3 Index]