sanitize_images {officer}	R Documentation

remove unused media from a document

Description

the function will scan the media directory and delete images that are not used anymore. This function is to be used when images have been replaced many times.

Usage

sanitize_images(x)

Arguments

x
rdocx or rpptx object

[Package officer version 0.4.1 Index]