set_autonum_bookmark {officer}	R Documentation

update bookmark of an autonumber run

Description

This function lets recycling a object made by run_autonum() by changing the bookmark value. This is useful to avoid calling run_autonum() several times because of many tables.

Usage

set_autonum_bookmark(x, bkm = NULL)

Arguments

x
an object of class run_autonum()

bkm
bookmark id to associate with autonumber run. Value can only be made of alpha numeric characters, ':', -' and '_'.

See Also

run_autonum()

Examples

z <- run_autonum(seq_id = "tab", pre_label = "Table ",
  bkm = "anytable")
set_autonum_bookmark(z, bkm = "anothertable")

[Package officer version 0.4.1 Index]