slip_in_column_break {officer}	R Documentation

add a column break

Description

add a column break into a Word document. A column break is used to add a break in a multi columns section in a Word Document.

This function will be deprecated in the next release because it is not efficient and make users write complex code, use run_columnbreak() instead.

Usage

slip_in_column_break(x, pos = "before")

Arguments

x
an rdocx object

pos
where to add the new element relative to the cursor, "after" or "before".

[Package officer version 0.4.1 Index]