slip_in_img {officer}	R Documentation

append an image

Description

append an image into a paragraph of an rdocx object.

This function will be deprecated in the next release because it is not efficient and make users write complex code. Use instead fpar() to build formatted paragraphs.

Usage

slip_in_img(x, src, style = NULL, width, height, pos = "after")

Arguments

x
an rdocx object

src
image filename, the basename of the file must not contain any blank.

style
text style

width
height in inches

height
height in inches

pos
where to add the new element relative to the cursor, "after" or "before".

[Package officer version 0.4.1 Index]