append seq field

append seq field into a paragraph of a docx object

slip_in_seqfield(x, str, style = "Normal", pos = "after")

Arguments

x

a docx device

str

seq field value

style

text style

pos

where to add the new element relative to the cursor, "after" or "before".

Examples

library(magrittr)
read_docx() %>%
  body_add_par("Time is: ", style = "Normal") %>%
  slip_in_seqfield(
    str = "TIME \u005C@ \"HH:mm:ss\" \u005C* MERGEFORMAT",
    style = 'strong') %>%

  body_add_par(" - This is a figure title", style = "centered") %>%
  slip_in_seqfield(str = "SEQ Figure \u005C* roman",
    style = 'Default Paragraph Font', pos = "before") %>%
  slip_in_text("Figure: ", style = "strong", pos = "before") %>%

  body_add_par(" - This is another figure title", style = "centered") %>%
  slip_in_seqfield(str = "SEQ Figure \u005C* roman",
    style = 'strong', pos = "before")  %>%
  slip_in_text("Figure: ", style = "strong", pos = "before") %>%
  body_add_par("This is a symbol: ", style = "Normal") %>%
  slip_in_seqfield(str = "SYMBOL 100 \u005Cf Wingdings",
    style = 'strong') %>%
  print(target = "seqfield.docx")

#> [1] "/Users/davidgohel/Github/officer/docs/reference/seqfield.docx"