slip_in_xml {officer}	R Documentation

add a wml string into a Word document

Description

The function add a wml string into the document after, before or on a cursor location.

This function will be deprecated in the next release because it is not efficient and make users write complex code. Use instead fpar() to build formatted paragraphs.

Usage

slip_in_xml(x, str, pos)

Arguments

x
an rdocx object

str
a wml string

pos
where to add the new element relative to the cursor, "after" or "before".

[Package officer version 0.4.1 Index]