sp_lineend {officer}	R Documentation

Line end properties

Description

Create a sp_lineend object that describes line end properties.

Usage

sp_lineend(type = "none", width = "med", length = "med")

## S3 method for class 'sp_lineend'
print(x, ...)

## S3 method for class 'sp_lineend'
update(object, type, width, length, ...)

Arguments

type
single character value specifying the line end type. Expected value is one of the following : default 'none' or 'triangle' or 'stealth' or 'diamond' or 'oval' or 'arrow'

width
single character value specifying the line end width Expected value is one of the following : default 'sm' or 'med' or 'lg'

length
single character value specifying the line end length Expected value is one of the following : default 'sm' or 'med' or 'lg'

x, object
sp_lineend object

...
further arguments - not used

Value

a sp_lineend object

See Also

sp_line

Other functions for defining shape properties: sp_line()

Examples

library(officer)

sp_lineend()
sp_lineend(type = "triangle")
sp_lineend(type = "arrow", width = "lg", length = "lg")
print(sp_lineend(type = "triangle", width = "lg"))
obj <- sp_lineend(type = "triangle", width = "lg")
update(obj, type = "arrow")

[Package officer version 0.7.3 Index]