table_layout {officer}	R Documentation

Algorithm for table layout

Description

When a table is displayed in a document, it can either be displayed using a fixed width or autofit layout algorithm:

fixed: uses fixed widths for columns. The width of the table is not changed regardless of the contents of the cells.

autofit: uses the contents of each cell and the table width to determine the final column widths.

Usage

table_layout(type = "autofit")

Arguments

type
'autofit' or 'fixed' algorithm. Default to 'autofit'.

See Also

Other functions for table definition: prop_table(), table_colwidths(), table_conditional_formatting(), table_stylenames(), table_width()

[Package officer version 0.4.1 Index]