table_width {officer}	R Documentation

Preferred width for a table

Description

Define the preferred width for a table.

Usage

table_width(width = 1, unit = "pct")

Arguments

width
value of the preferred width of the table.

unit
unit of the width. Possible values are 'in' (inches) and 'pct' (percent)

Word

All widths in a table are considered preferred because widths of columns can conflict and the table layout rules can require a preference to be overridden.

See Also

Other functions for table definition: prop_table(), table_colwidths(), table_conditional_formatting(), table_layout(), table_stylenames()

[Package officer version 0.4.1 Index]