unpack_folder {officer}	R Documentation

Extract files from a zip file

Description

Extract files from a zip file to a folder. The function returns the complete path to destination folder.

Usage

unpack_folder(file, folder)

Arguments

file
path of the archive to unzip

folder
folder to create

[Package officer version 0.4.1 Index]