Placeholder parameters annotation

generates a slide from each layout in the base document to identify the placeholder indexes, types, names, master names and layout names.

This is to be used when need to know what parameters should be used with ph_location* calls. The parameters are printed in their corresponding shapes.

Note that if there are duplicated ph_label, you should not use ph_location_label(). Hint: You can dedupe labels using layout_dedupe_ph_labels().

Usage

annotate_base(path = NULL, output_file = "annotated_layout.pptx")

Arguments

path

path to the pptx file to use as base document or NULL to use the officer default

output_file

filename to store the annotated powerpoint file or NULL to suppress generation

Value

rpptx object of the annotated PowerPoint file

See also

Examples

# To generate an anotation of the default base document with officer:
annotate_base(output_file = tempfile(fileext = ".pptx"))
#> pptx document with 7 slides
#> Available layouts and their associated master(s):
#>              layout       master
#> 1       Title Slide Office Theme
#> 2 Title and Content Office Theme
#> 3    Section Header Office Theme
#> 4       Two Content Office Theme
#> 5        Comparison Office Theme
#> 6        Title Only Office Theme
#> 7             Blank Office Theme

# To generate an annotation of the base document 'mydoc.pptx' and place the
# annotated output in 'mydoc_annotate.pptx'
# annotate_base(path = 'mydoc.pptx', output_file='mydoc_annotate.pptx')