create paragraph blocks

a list of blocks can be used to gather several blocks (paragraphs or tables) into a single object. The function is to be used when adding footnotes or formatted paragraphs into a new slide.

block_list(...)

Arguments

...
a list of objects of class fpar or flextable. When output is only for Word, objects of class external_img can also be used in fpar construction to mix text and images in a single paragraph.

See also

Examples

#' # block list ------

img.file <- file.path( R.home("doc"), "html", "logo.jpg" )
fpt_blue_bold <- fp_text(color = "#006699", bold = TRUE)
fpt_red_italic <- fp_text(color = "#C32900", italic = TRUE)

## This can be only be used in a MS word output as pptx does
## not support paragraphs made of text and images.
## (actually it can be used but image will not appear in the
## pptx output)
value <- block_list(
  fpar(ftext("hello world", fpt_blue_bold)),
  fpar(ftext("hello", fpt_blue_bold), " ",
       ftext("world", fpt_red_italic)),
  fpar(
    ftext("hello world", fpt_red_italic),
          external_img(
            src = img.file, height = 1.06, width = 1.39)))
value

#> [[1]]
#> $chunks
#> $chunks[[1]]
#> $value
#> [1] "hello world"
#> 
#> $pr
#>   size italic bold underlined   color     shading fontname vertical_align
#> 1   10  FALSE TRUE      FALSE #006699 transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "ftext" "cot"  
#> 
#> 
#> $fp_p
#>                     values
#> text.align            left
#> padding.top              0
#> padding.bottom           0
#> padding.left             0
#> padding.right            0
#> shading.color  transparent
#> borders:
#>        width color style
#> top        0 black solid
#> bottom     0 black solid
#> left       0 black solid
#> right      0 black solid
#> 
#> $fp_t
#>   size italic  bold underlined color     shading fontname vertical_align
#> 1   10  FALSE FALSE      FALSE black transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "fpar"
#> 
#> [[2]]
#> $chunks
#> $chunks[[1]]
#> $value
#> [1] "hello"
#> 
#> $pr
#>   size italic bold underlined   color     shading fontname vertical_align
#> 1   10  FALSE TRUE      FALSE #006699 transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "ftext" "cot"  
#> 
#> $chunks[[2]]
#> [1] " "
#> 
#> $chunks[[3]]
#> $value
#> [1] "world"
#> 
#> $pr
#>   size italic  bold underlined   color     shading fontname vertical_align
#> 1   10   TRUE FALSE      FALSE #C32900 transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "ftext" "cot"  
#> 
#> 
#> $fp_p
#>                     values
#> text.align            left
#> padding.top              0
#> padding.bottom           0
#> padding.left             0
#> padding.right            0
#> shading.color  transparent
#> borders:
#>        width color style
#> top        0 black solid
#> bottom     0 black solid
#> left       0 black solid
#> right      0 black solid
#> 
#> $fp_t
#>   size italic  bold underlined color     shading fontname vertical_align
#> 1   10  FALSE FALSE      FALSE black transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "fpar"
#> 
#> [[3]]
#> $chunks
#> $chunks[[1]]
#> $value
#> [1] "hello world"
#> 
#> $pr
#>   size italic  bold underlined   color     shading fontname vertical_align
#> 1   10   TRUE FALSE      FALSE #C32900 transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "ftext" "cot"  
#> 
#> $chunks[[2]]
#> [1] "/Library/Frameworks/R.framework/Resources/doc/html/logo.jpg"
#> attr(,"class")
#> [1] "external_img" "cot"         
#> attr(,"dims")
#> attr(,"dims")$width
#> [1] 1.39
#> 
#> attr(,"dims")$height
#> [1] 1.06
#> 
#> 
#> 
#> $fp_p
#>                     values
#> text.align            left
#> padding.top              0
#> padding.bottom           0
#> padding.left             0
#> padding.right            0
#> shading.color  transparent
#> borders:
#>        width color style
#> top        0 black solid
#> bottom     0 black solid
#> left       0 black solid
#> right      0 black solid
#> 
#> $fp_t
#>   size italic  bold underlined color     shading fontname vertical_align
#> 1   10  FALSE FALSE      FALSE black transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "fpar"
#> 
#> attr(,"class")
#> [1] "block_list"

doc <- read_docx()
doc <- body_add_blocks(doc, value)
print(doc, target = tempfile(fileext = ".docx"))

value <- block_list(
  fpar(ftext("hello world", fpt_blue_bold)),
  fpar(ftext("hello", fpt_blue_bold), " ",
       ftext("world", fpt_red_italic)),
  fpar(
    ftext("blah blah blah", fpt_red_italic)))
value

#> [[1]]
#> $chunks
#> $chunks[[1]]
#> $value
#> [1] "hello world"
#> 
#> $pr
#>   size italic bold underlined   color     shading fontname vertical_align
#> 1   10  FALSE TRUE      FALSE #006699 transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "ftext" "cot"  
#> 
#> 
#> $fp_p
#>                     values
#> text.align            left
#> padding.top              0
#> padding.bottom           0
#> padding.left             0
#> padding.right            0
#> shading.color  transparent
#> borders:
#>        width color style
#> top        0 black solid
#> bottom     0 black solid
#> left       0 black solid
#> right      0 black solid
#> 
#> $fp_t
#>   size italic  bold underlined color     shading fontname vertical_align
#> 1   10  FALSE FALSE      FALSE black transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "fpar"
#> 
#> [[2]]
#> $chunks
#> $chunks[[1]]
#> $value
#> [1] "hello"
#> 
#> $pr
#>   size italic bold underlined   color     shading fontname vertical_align
#> 1   10  FALSE TRUE      FALSE #006699 transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "ftext" "cot"  
#> 
#> $chunks[[2]]
#> [1] " "
#> 
#> $chunks[[3]]
#> $value
#> [1] "world"
#> 
#> $pr
#>   size italic  bold underlined   color     shading fontname vertical_align
#> 1   10   TRUE FALSE      FALSE #C32900 transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "ftext" "cot"  
#> 
#> 
#> $fp_p
#>                     values
#> text.align            left
#> padding.top              0
#> padding.bottom           0
#> padding.left             0
#> padding.right            0
#> shading.color  transparent
#> borders:
#>        width color style
#> top        0 black solid
#> bottom     0 black solid
#> left       0 black solid
#> right      0 black solid
#> 
#> $fp_t
#>   size italic  bold underlined color     shading fontname vertical_align
#> 1   10  FALSE FALSE      FALSE black transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "fpar"
#> 
#> [[3]]
#> $chunks
#> $chunks[[1]]
#> $value
#> [1] "blah blah blah"
#> 
#> $pr
#>   size italic  bold underlined   color     shading fontname vertical_align
#> 1   10   TRUE FALSE      FALSE #C32900 transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "ftext" "cot"  
#> 
#> 
#> $fp_p
#>                     values
#> text.align            left
#> padding.top              0
#> padding.bottom           0
#> padding.left             0
#> padding.right            0
#> shading.color  transparent
#> borders:
#>        width color style
#> top        0 black solid
#> bottom     0 black solid
#> left       0 black solid
#> right      0 black solid
#> 
#> $fp_t
#>   size italic  bold underlined color     shading fontname vertical_align
#> 1   10  FALSE FALSE      FALSE black transparent    Arial       baseline
#> 
#> attr(,"class")
#> [1] "fpar"
#> 
#> attr(,"class")
#> [1] "block_list"

doc <- read_pptx()
doc <- add_slide(doc)
doc <- ph_with(doc, value, location = ph_location_type(type = "body"))
print(doc, target = tempfile(fileext = ".pptx"))