Table of content for 'Word'

Create a representation of a table of content for Word documents.

Usage

block_toc(level = 3, style = NULL, seq_id = NULL, separator = ";")

Arguments

level

max title level of the table

style

optional. If not NULL, its value is used as style in the document that will be used to build entries of the TOC.

seq_id

optional. If not NULL, its value is used as sequence identifier in the document that will be used to build entries of the TOC. See also run_autonum() to specify a sequence identifier.

separator

unused, no effect

See also

Examples

block_toc(level = 2)
#> TOC - max level: 2
block_toc(style = "Table Caption")
#> TOC for style: Table Caption