docx_body_relationship {officer}	R Documentation

body xml document

Description

Get the body document as xml. This function is not to be used by end users, it has been implemented to allow other packages to work with officer.

Usage

docx_body_relationship(x)

Arguments

x
an rdocx object

Examples

doc <- read_docx()
docx_body_relationship(doc)

[Package officer version 0.4.1 Index]