Cell formatting properties

Create a fp_cell object that describes cell formatting properties.

fp_cell(border = fp_border(width = 0), border.bottom, border.left, border.top,
  border.right, vertical.align = "middle", margin = 0, margin.bottom,
  margin.top, margin.left, margin.right, background.color = "transparent",
  background.img.id = "rId1", background.img.src = NULL,
  text.direction = "lrtb", row_span = 1L, column_span = 1L)

# S3 method for fp_cell
format(x, type = "wml", ...)

# S3 method for fp_cell
print(x, ...)

# S3 method for fp_cell
update(object, border, border.bottom, border.left, border.top,
  border.right, vertical.align, margin = 0, margin.bottom, margin.top,
  margin.left, margin.right, background.color, background.img.id,
  background.img.src, text.direction, row_span, column_span, ...)

Arguments

border

shortcut for all borders.

border.bottom, border.left, border.top, border.right

fp_border for borders.

vertical.align

cell content vertical alignment - a single character value , expected value is one of "center" or "top" or "bottom"

margin

shortcut for all margins.

margin.bottom, margin.top, margin.left, margin.right

cell margins - 0 or positive integer value.

background.color

cell background color - a single character value specifying a valid color (e.g. "#000000" or "black").

background.img.id

id of the background image in the relation file.

background.img.src

path of the background image

text.direction

cell text rotation - a single character value, expected value is one of "lrtb", "tbrl", "btlr".

row_span, column_span

row and column span

x, object

object fp_cell

type

output type - one of 'wml', 'pml', 'html'.

...

further arguments - not used

Examples

obj <- fp_cell(margin = 1)
update( obj, margin.bottom = 5 )

#> border-bottom: 0pt solid rgba(0,0,0,1.00);border-top: 0pt solid rgba(0,0,0,1.00);border-left: 0pt solid rgba(0,0,0,1.00);border-right: 0pt solid rgba(0,0,0,1.00);background-color:transparent;margin-top:1pt;margin-bottom:5pt;margin-left:1pt;margin-right:1pt;vertical-align:middle;