fp_cell {officer}	R Documentation

Cell formatting properties

Description

Create a fp_cell object that describes cell formatting properties.

Usage

fp_cell(
  border = fp_border(width = 0),
  border.bottom,
  border.left,
  border.top,
  border.right,
  vertical.align = "center",
  margin = 0,
  margin.bottom,
  margin.top,
  margin.left,
  margin.right,
  background.color = "transparent",
  text.direction = "lrtb"
)

## S3 method for class 'fp_cell'
format(x, type = "wml", ...)

## S3 method for class 'fp_cell'
print(x, ...)

## S3 method for class 'fp_cell'
update(
  object,
  border,
  border.bottom,
  border.left,
  border.top,
  border.right,
  vertical.align,
  margin = 0,
  margin.bottom,
  margin.top,
  margin.left,
  margin.right,
  background.color,
  text.direction,
  ...
)

Arguments

border
shortcut for all borders.

border.bottom, border.left, border.top, border.right
fp_border for borders.

vertical.align
cell content vertical alignment - a single character value, expected value is one of "center" or "top" or "bottom"

margin
shortcut for all margins.

margin.bottom, margin.top, margin.left, margin.right
cell margins - 0 or positive integer value.

background.color
cell background color - a single character value specifying a valid color (e.g. "#000000" or "black").

text.direction
cell text rotation - a single character value, expected value is one of "lrtb", "tbrl", "btlr".

x, object
fp_cell object

type
output type - one of 'wml', 'pml', 'html'.

...
further arguments - not used

See Also

Other functions for defining formatting properties: fp_border(), fp_par(), fp_text()

Examples

obj <- fp_cell(margin = 1)
update( obj, margin.bottom = 5 )

[Package officer version 0.4.1 Index]