Tabulation mark properties object

create a tabulation mark properties setting object for Word or RTF. Results can be used as arguments of fp_tabs().

Once tabulation marks settings are defined, tabulation marks can be added with run_tab() inside a call to fpar() or with \t within 'flextable' content.

Usage

fp_tab(pos, style = "decimal")

Arguments

pos

Specifies the position of the tab stop (in inches).

style

style of the tab. Possible values are: "decimal", "left", "right" or "center".

See also

Examples

fp_tab(pos = 0.4, style = "decimal")
#> $pos
#> [1] 0.4
#> 
#> $style
#> [1] "decimal"
#> 
#> attr(,"class")
#> [1] "fp_tab"
fp_tab(pos = 1, style = "right")
#> $pos
#> [1] 1
#> 
#> $style
#> [1] "right"
#> 
#> attr(,"class")
#> [1] "fp_tab"