Tabs properties object

create a set of tabulation mark properties object for Word or RTF. Results can be used as arguments tabs of fp_par() and will only have effects in Word or RTF outputs.

Once a set of tabulation marks settings is defined, tabulation marks can be added with run_tab() inside a call to fpar() or with \t within 'flextable' content.

Usage

fp_tabs(...)

Arguments

...

fp_tab objects

See also

Examples

z <- fp_tabs(
  fp_tab(pos = 0.4, style = "decimal"),
  fp_tab(pos = 1, style = "decimal")
)
fpar(
  run_tab(), ftext("88."),
  run_tab(), ftext("987.45"),
  fp_p = fp_par(
    tabs = z
  )
)
#> $chunks
#> $chunks[[1]]
#> list()
#> attr(,"class")
#> [1] "run_tab" "run"    
#> 
#> $chunks[[2]]
#> text: 88.
#> format:
#> NULL
#> 
#> $chunks[[3]]
#> list()
#> attr(,"class")
#> [1] "run_tab" "run"    
#> 
#> $chunks[[4]]
#> text: 987.45
#> format:
#> NULL
#> 
#> 
#> $fp_p
#>                     values
#> text.align            left
#> padding.top              0
#> padding.bottom           0
#> padding.left             0
#> padding.right            0
#> shading.color  transparent
#> borders:
#>        width color style
#> top        0 black solid
#> bottom     0 black solid
#> left       0 black solid
#> right      0 black solid
#> 
#> $fp_t
#>   font.size italic bold underlined color shading fontname fontname_cs
#> 1        NA     NA   NA         NA    NA      NA       NA          NA
#>   fontname_eastasia fontname.hansi vertical_align
#> 1                NA             NA       baseline
#> 
#> attr(,"class")
#> [1] "fpar"  "block"