officer {officer}	R Documentation

officer: Manipulate Microsoft Word and PowerPoint Documents

Description

The officer package facilitates access to and manipulation of 'Microsoft Word' and 'Microsoft PowerPoint' documents from R.

Details

Examples of manipulations are:

read Word and PowerPoint files into data objects

add/edit/remove image, table and text content from documents and slides

write updated content back to Word and PowerPoint files

To learn more about officer, start with the vignettes: browseVignettes(package = "officer")

Author(s)

Maintainer: David Gohel david.gohel@ardata.fr

Other contributors:

Frank Hangler frank@plotandscatter.com (function body_replace_all_text) [contributor]

Liz Sander lsander@civisanalytics.com (several documentation fixes) [contributor]

Anton Victorson anton@victorson.se (fixes xml structures) [contributor]

Jon Calder jonmcalder@gmail.com (update vignettes) [contributor]

John Harrold john.m.harrold@gmail.com (function annotate_base) [contributor]

John Muschelli muschellij2@gmail.com (google doc compatibility) [contributor]

See Also

https://davidgohel.github.io/officer/

[Package officer version 0.4.1 Index]