page_size {officer}	R Documentation

page size object

Description

The function creates a representation of the dimensions of a page. The dimensions are defined by length, width and orientation. If the orientation is in landscape mode then the length becomes the width and the width becomes the length.

Usage

page_size(width = 21/2.54, height = 29.7/2.54, orient = "portrait")

Arguments

width, height
page width, page height (in inches).

orient
page orientation, either 'landscape', either 'portrait'.

See Also

Other functions for section definition: page_mar(), prop_section(), section_columns()

Examples

page_size(orient = "landscape")

[Package officer version 0.4.1 Index]