ph_add_text {officer}	R Documentation

append text

Description

append text in a placeholder. The function let you add text to an existing content in an exiisting shape, existing text will be preserved.

Usage

ph_add_text(
  x,
  str,
  type = "body",
  id = 1,
  id_chr = NULL,
  ph_label = NULL,
  style = fp_text(font.size = 0),
  pos = "after",
  href = NULL,
  slide_index = NULL
)

Arguments

x
an rpptx object

str
text to add

type
placeholder type

id
placeholder index (integer) for a duplicated type. This is to be used when a placeholder type is not unique in the layout of the current slide, e.g. two placeholders with type 'body'. To add onto the first, use id = 1 and id = 2 for the second one. Values can be read from slide_summary.

id_chr
deprecated.

ph_label
label associated to the placeholder. Use column ph_label of result returned by slide_summary.

style
text style, a fp_text object

pos
where to add the new element relative to the cursor, "after" or "before".

href
hyperlink to reach when clicking the text

slide_index
slide index to reach when clicking the text. It will be ignored if href is not NULL.

Usage

If your goal is to add formatted text in a new shape, use ph_with with a block_list instead of this function.

Note

This function will be deprecated in a next release because it is not efficient and make users write complex code. Use instead fpar() to build formatted paragraphs.

[Package officer version 0.4.1 Index]