plot_layout_properties {officer}	R Documentation

Plot slide layout properties

Description

Plot slide layout properties and print informations into defined placeholders. This can be useful to help visualise placeholders locations and identifier.

Usage

plot_layout_properties(x, layout = NULL, master = NULL, labels = TRUE)

Arguments

x
an rpptx object

layout
slide layout name to use

master
master layout name where layout is located

labels
if TRUE, placeholder labels will be printed, if FALSE placeholder types and identifiers will be printed.

See Also

Other functions for reading presentation informations: annotate_base(), color_scheme(), layout_properties(), layout_summary(), length.rpptx(), slide_size(), slide_summary()

Examples

x <- read_pptx()
plot_layout_properties( x = x, layout = "Title Slide",
  master = "Office Theme" )
plot_layout_properties( x = x, layout = "Two Content" )

[Package officer version 0.4.1 Index]