Table properties

Define table properties such as fixed or autofit layout, table width in the document, eventually column widths.

prop_table(
  style = NA_character_,
  layout = table_layout(),
  width = table_width(),
  stylenames = table_stylenames(),
  colwidths = table_colwidths(),
  tcf = table_conditional_formatting(),
  align = "center"
)

Arguments

style
table style to be used to format table

layout
layout defined by table_layout(),

width
table width in the document defined by table_width()

stylenames
columns styles defined by table_stylenames()

colwidths
column widths defined by table_colwidths()

tcf
conditional formatting settings defined by table_conditional_formatting()

align
table alignment (one of left, center or right)

See also

Examples

prop_table()

#> $style
#> [1] NA
#> 
#> $layout
#> $type
#> [1] "autofit"
#> 
#> attr(,"class")
#> [1] "table_layout"
#> 
#> $width
#> $width
#> [1] 1
#> 
#> $unit
#> [1] "pct"
#> 
#> attr(,"class")
#> [1] "table_width"
#> 
#> $colsizes
#> $widths
#> NULL
#> 
#> attr(,"class")
#> [1] "table_colwidths"
#> 
#> $stylenames
#> $stylenames
#> list()
#> 
#> attr(,"class")
#> [1] "table_stylenames"
#> 
#> $tcf
#> $first_row
#> [1] TRUE
#> 
#> $first_column
#> [1] FALSE
#> 
#> $last_row
#> [1] FALSE
#> 
#> $last_column
#> [1] FALSE
#> 
#> $no_hband
#> [1] FALSE
#> 
#> $no_vband
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "table_conditional_formatting"
#> 
#> $align
#> [1] "center"
#> 
#> attr(,"class")
#> [1] "prop_table"

to_wml(prop_table())

#> [1] "<w:tblPr>NA\"/><w:tblLayout w:type=\"autofit\"/><w:jc w:val=\"center\"/><w:tblLook w:firstRow=\"1\" w:lastRow=\"0\" w:firstColumn=\"0\" w:lastColumn=\"0\" w:noHBand=\"0\" w:noVBand=\"1\"/></w:tblPr>"