prop_table {officer}	R Documentation

Table properties

Description

Define table properties such as fixed or autofit layout, table width in the document, eventually column widths.

Usage

prop_table(
  style = NA_character_,
  layout = table_layout(),
  width = table_width(),
  stylenames = table_stylenames(),
  colwidths = table_colwidths(),
  tcf = table_conditional_formatting(),
  align = "center"
)

Arguments

style
table style to be used to format table

layout
layout defined by table_layout(),

width
table width in the document defined by table_width()

stylenames
columns styles defined by table_stylenames()

colwidths
column widths defined by table_colwidths()

tcf
conditional formatting settings defined by table_conditional_formatting()

align
table alignment (one of left, center or right)

See Also

Other functions for table definition: table_colwidths(), table_conditional_formatting(), table_layout(), table_stylenames(), table_width()

Examples

prop_table()
to_wml(prop_table())

[Package officer version 0.4.1 Index]