open a connexion to a 'Word' file

read and import a docx file as an R object representing the document.

read_docx(path = NULL)

# S3 method for docx
print(x, target = NULL, ...)

# S3 method for docx
length(x)

Arguments

path

path to the docx file to use a base document.

x

a docx object

target

path to the docx file to write

...

unused

Examples

# create a docx object with default template ---
read_docx()

#> docx document with 2 element(s)
#> Available styles are:
#>    style_type            style_id             style_name is_custom is_default
#> 1   paragraph              Normal                 Normal     FALSE       TRUE
#> 2   paragraph              Titre1              heading 1     FALSE      FALSE
#> 3   paragraph              Titre2              heading 2     FALSE      FALSE
#> 4   paragraph              Titre3              heading 3     FALSE      FALSE
#> 5   character      Policepardfaut Default Paragraph Font     FALSE       TRUE
#> 6       table       TableauNormal           Normal Table     FALSE       TRUE
#> 7   numbering         Aucuneliste                No List     FALSE       TRUE
#> 8   character              strong                 strong      TRUE      FALSE
#> 9   paragraph            centered               centered      TRUE      FALSE
#> 10      table              Grille             Table Grid     FALSE      FALSE
#> 11      table       tabletemplate         table_template      TRUE      FALSE
#> 12      table Listeclaire-Accent2    Light List Accent 2     FALSE      FALSE
#> 13  character           Titre1Car            Titre 1 Car      TRUE      FALSE
#> 14  character           Titre2Car            Titre 2 Car      TRUE      FALSE
#> 15  character           Titre3Car            Titre 3 Car      TRUE      FALSE

# write a docx object in a docx file ----
if( require(magrittr) ){
  read_docx() %>% print(target = "out.docx")
  # full path of produced file is returned
  print(.Last.value)
}

#> [1] "officer"   "stats"     "graphics"  "grDevices" "utils"     "datasets" 
#> [7] "methods"   "base"     

# how many element are there in the document ----
length( read_docx() )

#> [1] 2