Create an RTF document object

Creation of the object representing an RTF document which can then receive contents with the rtf_add() function and be written to a file with the print(x, target="doc.rtf") function.

Usage

rtf_doc(
  def_sec = prop_section(),
  normal_par = fp_par(),
  normal_chunk = fp_text(font.family = "Arial", font.size = 11)
)

Arguments

def_sec

a block_section object used to defined default section.

normal_par

an object generated by fp_par()

normal_chunk

an object generated by fp_text()

Value

an object of class rtf representing an empty RTF document.

See also

Examples

rtf_doc(normal_par = fp_par(padding = 3))
#> rtf document with 0 element(s)