auto number

Create an autonumbered chunk, i.e. a string representation of a sequence, each item will be numbered. These runs can also be bookmarked and be used later for cross references.

run_autonum(
  seq_id = "table",
  pre_label = "Table ",
  post_label = ": ",
  bkm = NULL,
  bkm_all = FALSE
)

Arguments

seq_id
sequence identifier

pre_label, post_label
text to add before and after number

bkm
bookmark id to associate with autonumber run. If NULL, no bookmark is added. Value can only be made of alpha numeric characters, '-' and '_'.

bkm_all
if TRUE, the bookmark will be set on the while string, if FALSE, the bookmark will be set on the number only. Default to FALSE. As an effect when a reference to this bookmark is used, the text can be like "Table 1" or "1" (pre_label is not included in the referenced text).

See also

Examples

run_autonum()

#> $seq_id
#> [1] "table"
#> 
#> $pre_label
#> [1] "Table "
#> 
#> $post_label
#> [1] ": "
#> 
#> $bookmark
#> NULL
#> 
#> $bookmark_all
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "run_autonum" "run"        

run_autonum(seq_id = "fig", pre_label = "fig. ")

#> $seq_id
#> [1] "fig"
#> 
#> $pre_label
#> [1] "fig. "
#> 
#> $post_label
#> [1] ": "
#> 
#> $bookmark
#> NULL
#> 
#> $bookmark_all
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "run_autonum" "run"        

run_autonum(seq_id = "tab", pre_label = "Table ", bkm = "anytable")

#> $seq_id
#> [1] "tab"
#> 
#> $pre_label
#> [1] "Table "
#> 
#> $post_label
#> [1] ": "
#> 
#> $bookmark
#> [1] "anytable"
#> 
#> $bookmark_all
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "run_autonum" "run"