Bookmark for 'Word'

Add a bookmark on a run object.

Usage

run_bookmark(bkm, run)

Arguments

bkm

bookmark id to associate with run. Value can only be made of alpha numeric characters, '-' and '_'.

run

a run object, made with a call to one of the "run functions for reporting".

usage

You can use this function in conjunction with fpar to create paragraphs consisting of differently formatted text parts. You can also use this function as an r chunk in an R Markdown document made with package officedown.

See also

Examples

ft <- fp_text(font.size = 12, bold = TRUE)
run_bookmark("par1", ftext("some text", ft))
#> $id
#> [1] "par1"
#> 
#> $run
#> $run[[1]]
#> text: some text
#> format:
#>   font.size italic bold underlined strike color     shading fontname
#> 1        12  FALSE TRUE      FALSE  FALSE black transparent    Arial
#>   fontname_cs fontname_eastasia fontname.hansi vertical_align
#> 1       Arial             Arial          Arial       baseline
#> 
#> 
#> attr(,"class")
#> [1] "run_bookmark" "run"