run_word_field {officer}	R Documentation

seqfield

Description

Create a Word computed field.

Usage

run_word_field(field, prop = NULL, seqfield = field)

run_seqfield(field, prop = NULL, seqfield = field)

Arguments

field, seqfield
computed field string (seqfield will be totally superseded by field in the future).

prop
formatting text properties returned by fp_text.

usage

You can use this function in conjunction with fpar to create paragraphs consisting of differently formatted text parts. You can also use this function as an r chunk in an R Markdown document made with package officedown.

Note

In the previous version, this function was called run_seqfield but the name was wrong and should have been run_word_field.

See Also

Other run functions for reporting: external_img(), ftext(), hyperlink_ftext(), run_autonum(), run_bookmark(), run_columnbreak(), run_footnoteref(), run_footnote(), run_linebreak(), run_pagebreak(), run_reference()

Other Word computed fields: run_autonum(), run_reference()

Examples

run_word_field(field = "PAGE  \\* MERGEFORMAT")
run_word_field(field = "Date \\@ \"MMMM d yyyy\"")

[Package officer version 0.4.1 Index]