set document properties

set Word or PowerPoint document properties. These are not visible in the document but are available as metadata of the document.

set_doc_properties(
  x,
  title = NULL,
  subject = NULL,
  creator = NULL,
  description = NULL,
  created = NULL
)

Arguments

x
an rdocx or rpptx object

title, subject, creator, description
text fields

created
a date object

Note

The "last modified" and "last modified by" fields will be automatically be updated when the file is written.

See also

Examples

x <- read_docx()
x <- set_doc_properties(x, title = "title",
  subject = "document subject", creator = "Me me me",
  description = "this document is empty",
  created = Sys.time())
x <- doc_properties(x)