set_doc_properties {officer}	R Documentation

set document properties

Description

set Word or PowerPoint document properties. These are not visible in the document but are available as metadata of the document.

Usage

set_doc_properties(
  x,
  title = NULL,
  subject = NULL,
  creator = NULL,
  description = NULL,
  created = NULL
)

Arguments

x
an rdocx or rpptx object

title, subject, creator, description
text fields

created
a date object

Note

The "last modified" and "last modified by" fields will be automatically be updated when the file is written.

See Also

Other functions for Word document informations: doc_properties(), docx_bookmarks(), docx_dim(), length.rdocx(), styles_info()

Examples

x <- read_docx()
x <- set_doc_properties(x, title = "title",
  subject = "document subject", creator = "Me me me",
  description = "this document is empty",
  created = Sys.time())
x <- doc_properties(x)

[Package officer version 0.4.1 Index]