get PowerPoint slide content in a data.frame

get content and positions of current slide into a data.frame. Data for any tables, images, or paragraphs are imported into the resulting data.frame.

slide_summary(x, index = NULL)

Arguments

x
an rpptx object

index
slide index

Note

The column id of the result is not to be used by users. This is a technical string id whose value will be used by office when the document will be rendered. This is not related to argument index required by functions ph_with.

See also

Examples

my_pres <- read_pptx()
my_pres <- add_slide(my_pres)
my_pres <- ph_with(my_pres, format(Sys.Date()),
  location = ph_location_type(type="dt"))
my_pres <- add_slide(my_pres)
my_pres <- ph_with(my_pres, iris[1:2,],
  location = ph_location_type(type="body"))
slide_summary(my_pres)

#>   type id              ph_label offx offy cx cy
#> 1 body  2 Content Placeholder 2   NA   NA NA NA
#>                                                                                                                              text
#> 1 {5C22544A-7EE6-4342-B048-85BDC9FD1C3A}Sepal.LengthSepal.WidthPetal.LengthPetal.WidthSpecies5.13.51.40.2setosa4.93.01.40.2setosa

slide_summary(my_pres, index = 1)

#>   type id           ph_label offx     offy       cx        cy       text
#> 1   dt  2 Date Placeholder 3  0.5 6.951389 2.333333 0.3993056 2020-06-25