slip_in_footnote {officer}	R Documentation

append a footnote

Description

append a new footnote into a paragraph of an rdocx object

Usage

slip_in_footnote(x, style = NULL, blocks, pos = "after")

Arguments

x
an rdocx object

style
text style to be used for the reference note

blocks
set of blocks to be used as footnote content returned by function block_list.

pos
where to add the new element relative to the cursor, "after" or "before".

Note

This function will be deprecated in the next release because it is not efficient and make users write complex code. Use instead fpar() to build formatted paragraphs.

[Package officer version 0.4.1 Index]